Release 31 / 24-0D

+++ opensyde_tool/libs/opensyde_core/project/system/C_OscSystemDefinitionFiler.cpp
@@ -326,8 +326,8 @@ int32_t C_OscSystemDefinitionFiler::h_LoadNodes(std::vector<C_OscNode> & orc_Nod
       if (u32_ExpectedSize != orc_Nodes.size())
       {
          C_SclString c_Tmp;
-         c_Tmp.PrintFormatted("Unexpected nodes count, expected: %i, got %i", u32_ExpectedSize,
-                              orc_Nodes.size());
+         c_Tmp.PrintFormatted("Unexpected nodes count, expected: %u, got %u", u32_ExpectedSize,
+                              static_cast<uint32_t>(orc_Nodes.size()));
          osc_write_log_warning("Load file", c_Tmp.c_str());
       }
    }
@@ -426,8 +426,8 @@ int32_t C_OscSystemDefinitionFiler::h_LoadBuses(std::vector<C_OscSystemBus> & or
       if (u32_ExpectedSize != orc_Buses.size())
       {
          C_SclString c_Tmp;
-         c_Tmp.PrintFormatted("Unexpected bus count, expected: %i, got %i", u32_ExpectedSize,
-                              orc_Buses.size());
+         c_Tmp.PrintFormatted("Unexpected bus count, expected: %u, got %u", u32_ExpectedSize,
+                              static_cast<uint32_t>(orc_Buses.size()));
          osc_write_log_warning("Load file", c_Tmp.c_str());
       }
    }
